+++ okhttp-tls/src/main/kotlin/okhttp3/tls/HeldCertificate.kt
@@ -24,8 +24,6 @@ import java.security.PrivateKey
 import java.security.PublicKey
 import java.security.SecureRandom
 import java.security.Security
-import java.security.cert.CertificateFactory
-import java.security.cert.CertificateParsingException
 import java.security.cert.X509Certificate
 import java.security.interfaces.ECPublicKey
 import java.security.interfaces.RSAPrivateKey
@@ -36,7 +34,6 @@ import java.util.UUID
 import java.util.concurrent.TimeUnit
 import javax.security.auth.x500.X500Principal
 import okhttp3.internal.canParseAsIpAddress
-import okio.Buffer
 import okio.ByteString
 import okio.ByteString.Companion.decodeBase64
 import okio.ByteString.Companion.toByteString
@@ -120,6 +117,7 @@ import org.bouncycastle.x509.X509V3CertificateGenerator
  * a chain of certificates. The server uses a set of trusted root certificates to authenticate the
  * client. Subject alternative names are not used for client authentication.
  */
+@Suppress("DEPRECATION")
 class HeldCertificate(
   @get:JvmName("keyPair") val keyPair: KeyPair,
   @get:JvmName("certificate") val certificate: X509Certificate
@@ -490,67 +488,31 @@ class HeldCertificate(
     }
 
     private fun decode(certificatePem: String, pkcs8Base64Text: String): HeldCertificate {
-      val certificate = try {
-        decodeCertificate(certificatePem)
-      } catch (e: GeneralSecurityException) {
-        throw IllegalArgumentException("failed to decode certificate", e)
+      val certificate = certificatePem.decodeCertificatePem()
+
+      val pkcs8Bytes = pkcs8Base64Text.decodeBase64()
+          ?: throw IllegalArgumentException("failed to decode private key")
+
+      // The private key doesn't tell us its type but it's okay because the certificate knows!
+      val keyType = when (certificate.publicKey) {
+        is ECPublicKey -> "EC"
+        is RSAPublicKey -> "RSA"
+        else -> throw IllegalArgumentException("unexpected key type: ${certificate.publicKey}")
       }
 
-      val privateKey = try {
-        val pkcs8Bytes = pkcs8Base64Text.decodeBase64()
-            ?: throw IllegalArgumentException("failed to decode private key")
-
-        // The private key doesn't tell us its type but it's okay because the certificate knows!
-        val keyType = when (certificate.publicKey) {
-          is ECPublicKey -> "EC"
-          is RSAPublicKey -> "RSA"
-          else -> throw IllegalArgumentException("unexpected key type: ${certificate.publicKey}")
-        }
-
-        decodePkcs8(pkcs8Bytes, keyType)
-      } catch (e: GeneralSecurityException) {
-        throw IllegalArgumentException("failed to decode private key", e)
-      }
+      val privateKey = decodePkcs8(pkcs8Bytes, keyType)
 
       val keyPair = KeyPair(certificate.publicKey, privateKey)
       return HeldCertificate(keyPair, certificate)
     }
 
-    /**
-     * Decodes a multiline string that contains a [certificate][certificatePem] which is
-     * [PEM-encoded][rfc_7468]. A typical input string looks like this:
-     *
-     * ```
-     * -----BEGIN CERTIFICATE-----
-     * MIIBYTCCAQegAwIBAgIBKjAKBggqhkjOPQQDAjApMRQwEgYDVQQLEwtlbmdpbmVl
-     * cmluZzERMA8GA1UEAxMIY2FzaC5hcHAwHhcNNzAwMTAxMDAwMDA1WhcNNzAwMTAx
-     * MDAwMDEwWjApMRQwEgYDVQQLEwtlbmdpbmVlcmluZzERMA8GA1UEAxMIY2FzaC5h
-     * cHAwWTATBgcqhkjOPQIBBggqhkjOPQMBBwNCAASda8ChkQXxGELnrV/oBnIAx3dD
-     * ocUOJfdz4pOJTP6dVQB9U3UBiW5uSX/MoOD0LL5zG3bVyL3Y6pDwKuYvfLNhoyAw
-     * HjAcBgNVHREBAf8EEjAQhwQBAQEBgghjYXNoLmFwcDAKBggqhkjOPQQDAgNIADBF
-     * AiAyHHg1N6YDDQiY920+cnI5XSZwEGhAtb9PYWO8bLmkcQIhAI2CfEZf3V/obmdT
-     * yyaoEufLKVXhrTQhRfodTeigi4RX
-     * -----END CERTIFICATE-----
-     * ```
-     */
-    @JvmStatic
-    fun decodeCertificate(pem: String): X509Certificate {
-      val certificateFactory = CertificateFactory.getInstance("X.509")
-      val certificates = certificateFactory
-          .generateCertificates(Buffer().writeUtf8(pem).inputStream())
-
-      try {
-        return certificates.single() as X509Certificate
-      } catch (nsee: NoSuchElementException) {
-        throw CertificateParsingException(nsee)
-      } catch (iae: IllegalArgumentException) {
-        throw CertificateParsingException(iae)
-      }
-    }
-
     private fun decodePkcs8(data: ByteString, keyAlgorithm: String): PrivateKey {
-      val keyFactory = KeyFactory.getInstance(keyAlgorithm)
-      return keyFactory.generatePrivate(PKCS8EncodedKeySpec(data.toByteArray()))
+      try {
+        val keyFactory = KeyFactory.getInstance(keyAlgorithm)
+        return keyFactory.generatePrivate(PKCS8EncodedKeySpec(data.toByteArray()))
+      } catch (e: GeneralSecurityException) {
+        throw IllegalArgumentException("failed to decode private key", e)
+      }
     }
   }
 }
